Enhance report services with date filtering and type casting
- Added a date range filter to the average days to close calculation in CustomersReportService to refine data accuracy.
- Updated type casting in PlatformReportService to ensure consistent float conversion when summing expense amounts.

These changes improve the precision of report data and maintain type integrity across calculations.

+++ app/Domain/Reports/Services/PlatformReportService.php
@@ -416,7 +416,7 @@ final class PlatformReportService
     {
         try {
             return (float) round(
-                DB::table('rm_expenses')
+                (float) DB::table('rm_expenses')
                     ->where('user_id', $userId)
                     ->whereBetween('created_at', [$start, $end])
                     ->sum('amount_value'),
